+++ ft_calloc.c
@@ -6,13 +6,13 @@
 /*   By: atereso- <atereso-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/28 16:17:04 by atereso-          #+#    #+#             */
-/*   Updated: 2021/11/16 18:54:29 by atereso-         ###   ########.fr       */
+/*   Updated: 2021/11/25 16:57:06 by atereso-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_calloc(unsigned int count, unsigned int size)
+void	*ft_calloc(size_t count, size_t size)
 {
 	char	*allocator;
 	int		i;
